+++ app/src/main/java/com/srpf/srpf_gp_2/Activity/SignUpActivity.java
@@ -22,6 +22,7 @@ import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.srpf.srpf_gp_2.R;
+import com.srpf.srpf_gp_2.Utils.SharedPrefManager;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -33,20 +34,24 @@ import java.util.Map;
 
 public class SignUpActivity extends AppCompatActivity implements AdapterView.OnItemSelectedListener {
     private static final String url = "https://oneclickhub.in/srpf/userRegister.php";
-    EditText sname, semail, smobile, sdgpno, spass;
-    Button btn;
+    private EditText sname, semail, smobile, sdgpno, spass;
+    private Button btn;
     private Spinner sp1, sp2;
     private ArrayList<String> countryList = new ArrayList<>();
     private ArrayList<String> cityList = new ArrayList<>();
     private ArrayAdapter<String> countryAdapter;
     private ArrayAdapter<String> cityAdapter;
     private RequestQueue requestQueue;
+    private String firebaseToken;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_singup);
         requestQueue = Volley.newRequestQueue(this);
+
+        firebaseToken = SharedPrefManager.getInstance(this).getDeviceToken();
+
         sp1 = findViewById(R.id.sp1);
         sp2 = findViewById(R.id.sp2);
         sname = (EditText) findViewById(R.id.sname);
@@ -60,8 +65,12 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
         btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                register_user(sname.getText().toString(), semail.getText().toString(), smobile.getText().toString(), sdgpno.getText().toString(), spass.getText().toString(),
-                        sp1.getSelectedItem().toString(), sp2.getSelectedItem().toString());
+                register_user(sname.getText().toString(),
+                        semail.getText().toString(),
+                        smobile.getText().toString(),
+                        sdgpno.getText().toString(),
+                        spass.getText().toString(),
+                        sp1.getSelectedItem().toString());
             }
         });
 
@@ -100,7 +109,6 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
 
     @Override
     public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
-
         if (adapterView.getId() == R.id.sp1) {
             cityList.clear();
             String selectedCountry = adapterView.getSelectedItem().toString();
@@ -112,7 +120,6 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
                 @Override
                 public void onResponse(JSONObject response) {
                     try {
-
                         JSONArray jsonArray = response.getJSONArray("tbl_company");
                         for (int i = 0; i < jsonArray.length(); i++) {
                             JSONObject jsonObject = jsonArray.getJSONObject(i);
@@ -135,11 +142,10 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
                 }
             });
             requestQueue.add(jsonObjectRequest);
-
         }
     }
 
-    public void register_user(final String name, final String email, final String mobno, final String dgpno, final String pass, final String spn1, String spn2) {
+    public void register_user(final String name, final String email, final String mobno, final String dgpno, final String pass, final String spn1) {
         StringRequest request = new StringRequest(Request.Method.POST, url, new Response.Listener<String>() {
             @Override
             public void onResponse(String response) {
@@ -148,8 +154,6 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
                 smobile.setText("");
                 sdgpno.setText("");
                 spass.setText("");
-
-
                 Toast.makeText(getApplicationContext(), response, Toast.LENGTH_LONG).show();
             }
         }, new Response.ErrorListener() {
@@ -160,10 +164,7 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
                 smobile.setText("");
                 sdgpno.setText("");
                 spass.setText("");
-
-
                 Toast.makeText(getApplicationContext(), error.toString(), Toast.LENGTH_LONG).show();
-
             }
         }
         ) {
@@ -177,7 +178,8 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
                 map.put("DGP_No", dgpno);
                 map.put("password", pass);
                 map.put("CompanyName", spn1);
-                map.put("CompanyID", spn2);
+                map.put("Token", firebaseToken);
+                map.put("CompanyID", spn1);
                 return map;
             }
         };
@@ -187,13 +189,12 @@ public class SignUpActivity extends AppCompatActivity implements AdapterView.OnI
     }
 
     public void btnLogin(View view) {
-        Intent intent = new Intent(this, MainActivity.class);
+        Intent intent = new Intent(this, LoginActivity.class);
         startActivity(intent);
         finish();
     }
 
     @Override
     public void onNothingSelected(AdapterView<?> adapterView) {
-
     }
 }
