+++ app/src/main/java/com/srpf/srpf_gp_2/Services/MyFirebaseInstanceIDService.java
@@ -1,15 +1,24 @@
 package com.srpf.srpf_gp_2.Services;
 
+import android.app.PendingIntent;
+import android.content.Intent;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.NotificationManagerCompat;
 
 import com.google.firebase.messaging.FirebaseMessagingService;
+import com.google.firebase.messaging.RemoteMessage;
+import com.srpf.srpf_gp_2.Activity.DashboardActivity;
+import com.srpf.srpf_gp_2.Activity.LoginActivity;
+import com.srpf.srpf_gp_2.R;
 import com.srpf.srpf_gp_2.Utils.SharedPrefManager;
 
 public class MyFirebaseInstanceIDService extends FirebaseMessagingService {
 
-    private static final String TAG = "MyFirebaseIIDService";
+    private static final String TAG = "PushNotification";
+    private static final String CHANNEL_ID = "101";
 
     @Override
     public void onNewToken(@NonNull String refreshedToken) {
@@ -18,8 +27,33 @@ public class MyFirebaseInstanceIDService extends FirebaseMessagingService {
         storeToken(refreshedToken);
     }
 
+    @Override
+    public void onMessageReceived(@NonNull RemoteMessage remoteMessage) {
+        super.onMessageReceived(remoteMessage);
+        showNotification(remoteMessage.getNotification().getTitle(), remoteMessage.getNotification().getBody());
+    }
+
     private void storeToken(String token) {
-        //saving the token on shared preferences
         SharedPrefManager.getInstance(getApplicationContext()).saveDeviceToken(token);
+    }
+
+    private void showNotification(String title, String message) {
+        Intent intent = new Intent(this, DashboardActivity.class);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK );
+        PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_MUTABLE);
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
+                .setSmallIcon(R.drawable.ic_launcher_background)
+                .setContentTitle(title)
+                .setContentText(message)
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+                // Set the intent that will fire when the user taps the notification
+                .setContentIntent(pendingIntent)
+                .setAutoCancel(true);
+
+        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(this);
+
+        // notificationId is a unique int for each notification that you must define
+        notificationManager.notify(1, builder.build());
     }
 }
