+++ app/src/main/res/layout/activity_main.xml
@@ -1,15 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
     style="@style/loginParent"
-    tools:context=".Activity.MainActivity">
-    <ScrollView
-        style="@style/parent">
+    android:orientation="vertical"
+    tools:context=".Activity.LoginActivity">
 
+    <ScrollView style="@style/parent">
         <RelativeLayout style="@style/parent">
-
             <LinearLayout
                 style="@style/parent"
                 android:layout_width="match_parent"
@@ -21,12 +18,9 @@
 
                 <ImageView
                     style="@style/parent"
-
                     android:background="@drawable/ic_login_bk"
                     android:contentDescription="@string/login_background" />
-
             </LinearLayout>
-
 
             <LinearLayout
                 android:layout_width="match_parent"
@@ -37,12 +31,8 @@
                 android:layout_marginRight="30dp"
                 android:orientation="vertical">
 
-
                 <include layout="@layout/activity_loginform" />
             </LinearLayout>
-
         </RelativeLayout>
-
     </ScrollView>
-
-    </RelativeLayout>
+</RelativeLayout>
