+++ app/src/main/res/layout/singupform.xml
@@ -3,11 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/loginCard"
-    android:orientation="vertical"
-    android:background="@color/whiteCardColor"
+    android:layout_gravity="center"
     android:layout_marginTop="30dp"
     android:layout_marginBottom="@dimen/loginViewsMargin"
-    android:layout_gravity="center"
+    android:background="@color/whiteCardColor"
+    android:orientation="vertical"
     app:cardCornerRadius="@dimen/loginCardRadius"
     tools:context=".Activity.SignUpActivity">
 
@@ -36,11 +36,9 @@
             android:autofillHints=""
             android:ems="10"
             android:hint="Enter Your Name"
-            android:textSize="20sp"
             android:minHeight="48dp"
-            android:textColorHint="#757575"        />
-
-
+            android:textColorHint="#757575"
+            android:textSize="20sp" />
 
         <EditText
             android:id="@+id/smobile"
@@ -62,29 +60,29 @@
             android:hint="Enter Email Id."
             android:minHeight="48dp"
             android:textColorHint="#757575" />
+
         <Spinner
             android:id="@+id/sp1"
+            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.ExposedDropdownMenu"
             android:layout_width="350dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="5dp"
-            android:textSize="20sp"
-            android:ems="10"
             android:background="@drawable/bg_spinner_fruit"
             android:dropDownSelector="@color/colorWhite"
+            android:ems="10"
             android:minHeight="48dp"
             android:textColorHint="#757575"
-            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.ExposedDropdownMenu"/>
+            android:textSize="20sp" />
 
         <Spinner
             android:id="@+id/sp2"
             android:layout_width="350dp"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:visibility="gone"
-            android:textColorHint="#757575"
             android:ems="10"
             android:minHeight="48dp"
-            />
+            android:textColorHint="#757575"
+            android:visibility="gone" />
 
         <EditText
             android:id="@+id/sdgpno"
@@ -95,7 +93,7 @@
             android:ems="10"
             android:hint="Enter DGP No."
             android:minHeight="48dp"
-            android:textColorHint="#757575"/>
+            android:textColorHint="#757575" />
 
         <EditText
             android:id="@+id/spass"
@@ -107,7 +105,7 @@
             android:hint="Enter_password"
             android:inputType="textPassword"
             android:minHeight="48dp"
-            android:textColorHint="#757575"/>
+            android:textColorHint="#757575" />
 
         <Button
             android:id="@+id/btnreg"
@@ -116,7 +114,6 @@
             android:layout_marginTop="10dp"
             android:text="Register" />
 
-
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -124,21 +121,18 @@
             android:layout_marginTop="10dp"
             android:minHeight="48dp"
             android:onClick="btnLogin"
-            android:textColor="@color/colorPrimaryDark"
             android:text="Already registered login here"
-            android:textAlignment="center" />
-
+            android:textAlignment="center"
+            android:textColor="@color/colorPrimaryDark" />
 
         <TextView
             android:id="@+id/tv"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_gravity="center"
             android:layout_marginTop="5dp"
-            android:textSize="20sp"
-            android:textStyle="bold"
             android:text="@string/create_new_account"
-            android:layout_gravity="center"/>
-
+            android:textSize="20sp"
+            android:textStyle="bold" />
     </LinearLayout>
-
 </androidx.cardview.widget.CardView>
